ver.0.36.f.1 : V-AUS debugged.
Introducing a new I/O file: aus-restart file.
Information of atomic coordinates including exceeding PBC boundaries is recorded.
If the aus-restart file is not specified,
the center of mass of each group is calculated from the initial coordinates without any revision of PBC.
In other words, when a group is on a boundary, the center of mass becomes non-sense.
The initial coordinates should be prepared carefully.

The file format of .cls file is updated as "v.0.36.f".