Output warnings in the mdinput_generator.py
For the first run of AUS, the atom group should not be across the cell boundary.
The minimum distance between the center and group atoms are checked in the script.

Warning messages should be appended into  the list MDInputGen.msg_warnings.
The messages is output into the file (WARNING.celeste as default)